+++ Server/TCPLib.Server/Comands/Ban.cs
@@ -21,7 +21,7 @@ namespace TCPLib.Server.Commands
                 return false;
             }
             string ip = "";
-            NetClient client = null;
+            Client client = null;
             if (args[0].IndexOf('.') == -1)
             {
                 var clientlist = Client.clients.Where(x => x.id.ToString() == args[0]);
