+++ Server/TCPLib.Server/Net/Client/Client.cs
@@ -9,25 +9,25 @@ using System.Threading;
 
 namespace TCPLib.Server.Net
 {
-    public abstract partial class NetClient : IDisposable
+    public partial class Client : IDisposable
     {
         public EncryptType EncryptType { get; set; } = EncryptType.RSA;
 
         public TcpClient client { get; set; }
         public NetworkStream stream { get; set; }
         public uint id { get; private set; }
-        public static IReadOnlyCollection<NetClient> clients
+        public static IReadOnlyCollection<Client> clients
         {
             get
             {
                 return _clients;
             }
         }
-        protected static List<NetClient> _clients = new List<NetClient>();
+        protected static List<Client> _clients = new List<Client>();
         public Encryptor Encryptor { get; set; }
 
         protected CancellationTokenSource OnKick;
-        public NetClient(TcpClient client, NetworkStream stream)
+        public Client(TcpClient client, NetworkStream stream)
         {
             this.client = client;
             if (clients.Count == 0) id = 0;
@@ -64,6 +64,16 @@ namespace TCPLib.Server.Net
             {
                 OnDisconnected();
             }
+        }
+        public void Ban(string Reason = "", TimeSpan? time = null)
+        {
+            var list = SaveFiles.Ban.Load().ToList();
+            if (time is null)
+                list.Add(SaveFiles.Ban.CreateBan(this, Reason));
+            else
+                list.Add(SaveFiles.Ban.CreateBan(this, Reason, DateTime.UtcNow + time));
+            SaveFiles.Ban.Save(list.ToArray());
+            GC.Collect();
         }
     }
     public enum EncryptType
